+++ src/mig/logging_posix.go
@@ -33,7 +33,7 @@ type Logging struct {
 
 // Log defines a log entry
 type Log struct {
-	OpID, ActionID, CommandID float64
+	OpID, ActionID, CommandID uint64
 	Sev, Desc                 string
 	Priority                  syslog.Priority
 }
