+++ src/mig/logging_windows.go
@@ -34,7 +34,7 @@ type Logging struct {
 
 // Log defines a log entry
 type Log struct {
-	OpID, ActionID, CommandID float64
+	OpID, ActionID, CommandID uint64
 	Sev, Desc                 string
 	Priority                  int
 }
